requirements.txt
Pillow>=9.4.0
pascal-voc-writer>=0.1.4

